project/requirements.txt
playwright
pytest
pytest-playwright
requests
plantuml
openpyxl
allure-pytest
python-dotenv
playwright==1.44.0
pytest==7.4.0
pytest-playwright==0.3.0
python-dotenv==1.0.0
